separated the tracks classes from the tracksClass.py file. Now, any class ending in Track.py that is placed in the tracks folder would be interpreted as a track class
