+++ pygenometracks/tracks/BedTrack.py
@@ -1,0 +1,584 @@
+from . GenomeTrack import GenomeTrack
+from .. readBed import ReadBed
+from matplotlib.patches import Rectangle
+import matplotlib
+from .. utilities import opener
+from intervaltree import IntervalTree, Interval
+import numpy as np
+
+DEFAULT_BED_COLOR = '#1f78b4'
+
+
+class BedTrack(GenomeTrack):
+    SUPPORTED_ENDINGS = ['bed', 'bed3', 'bed6', 'bed12', 'bed.gz', 'bed3.gz', 'bed6.gz', 'bed12.gz']
+    TRACK_TYPE = 'bed'
+    OPTIONS_TXT = GenomeTrack.OPTIONS_TXT + """
+# if the type=genes is given
+# the the file is interpreted as gene
+# file. If the bed file contains the exon
+# structure (bed 12) then this is plotted. Otherwise
+# a region **with direction** is plotted.
+# if the bed file contains a column for color (column 9), then this color can be used by
+# setting:
+# color = bed_rgb
+color = darkblue
+#if color is a valid colormap name (like RbBlGn), then the score is mapped
+# to the colormap. If the color is simply a color name, then this color is used and the score is not considered.
+# For the colormap option, the the min_value and max_value for the score can be provided, otherwise
+# the maximum score and minimum score found are used.
+#color = RdYlBu
+#min_value=0
+#max_value=100
+# height of track in cm
+height = 5
+# to turn off/on printing of labels
+labels = off
+# optional: font size can be given to override the default size
+fontsize = 10
+# optional: line width
+#line width = 0.5
+# the display parameter defines how the bed file is plotted.
+# The options are ['colapsed', 'interleaved', 'triangles'] This options asume that the regions do not overlap.
+# `collapsed`: The bed regions are plotted one after the other in one line.
+# `interleaved`: The bed regions are plotted in two lines, first up, then down, then up etc.
+# if display is not given, then each region is plotted using the gene style
+#optional, default is black. To remove the background color, simply set 'color' and 'background color' to the
+# same value
+#border color = black
+# style to plot the genes when they have exon information
+#style = UCSC
+#style = flybase
+# maximum number of gene rows to be plotted. This
+# field is useful to limit large number of close genes
+# to be printed over many rows. When several images want
+# to be combined this must be set to get equal size, otherwise, on each image the height of each gene changes
+#gene rows = 10
+# if the track wants to be plotted on top of the previous track. Options are 'yes' or 'share-y'. For the 'share-y'
+# option the y axis values is shared between this plot and the overlay plot. Otherwise, each plot use its own scale
+#overlay previous = yes
+# by default the ymax is the number of
+# rows occupied by the genes in the region plotted. However,
+# by setting this option, the global maximum is used instead.
+# This is useful to combine images that are all consistent and
+# have the same number of rows.
+#global max row = yes
+# optional. If not given is guessed from the file ending.
+file_type = {}
+    """.format(TRACK_TYPE)
+
+    def __init__(self, *args, **kwarg):
+        super(BedTrack, self).__init__(*args, **kwarg)
+        self.bed_type = None  # once the bed file is read, this is bed3, bed6 or bed12
+        self.len_w = None  # this is the length of the letter 'w' given the font size
+        self.interval_tree = {}  # interval tree of the bed regions
+
+        from matplotlib import font_manager
+        if 'fontsize' not in self.properties:
+            self.properties['fontsize'] = 12
+        else:
+            self.properties['fontsize'] = float(self.properties['fontsize'])
+
+        self.fp = font_manager.FontProperties(size=self.properties['fontsize'])
+
+        if 'color' not in self.properties:
+            self.properties['color'] = DEFAULT_BED_COLOR
+        if 'border color' not in self.properties:
+            self.properties['border color'] = 'black'
+        if 'labels' not in self.properties:
+            self.properties['labels'] = 'on'
+        if 'style' not in self.properties:
+            self.properties['style'] = 'flybase'
+        if 'display' not in self.properties:
+            self.properties['display'] = 'stacked'
+        if 'interval height' not in self.properties:
+            self.properties['interval_height'] = 100
+        if 'line width' not in self.properties:
+            self.properties['line width'] = 0.5
+
+        self.colormap = None
+
+        # check if the color given is a color map
+        if not matplotlib.colors.is_color_like(self.properties['color']) and self.properties['color'] != 'bed_rgb':
+            # check if the color is a valid colormap name
+            if self.properties['color'] not in matplotlib.cm.datad:
+                self.log.warning("*WARNING* color: '{}' for section {} is not valid. Color has "
+                            "been set to {}".format(self.properties['color'], self.properties['section_name'],
+                                                    DEFAULT_BED_COLOR))
+                self.properties['color'] = DEFAULT_BED_COLOR
+            else:
+                self.colormap = self.properties['color']
+
+        # to set the distance between rows
+        self.row_scale = self.properties['interval_height'] * 2.3
+
+        self.interval_tree, min_score, max_score = self.process_bed()
+        if self.colormap is not None:
+            if 'min_value' in self.properties:
+                min_score = self.properties['min_value']
+            if 'max_value' in self.properties:
+                max_score = self.properties['max_value']
+
+            norm = matplotlib.colors.Normalize(vmin=min_score,
+                                               vmax=max_score)
+
+            cmap = matplotlib.cm.get_cmap(self.properties['color'])
+            self.colormap = matplotlib.cm.ScalarMappable(norm=norm, cmap=cmap)
+
+    def get_length_w(self, fig_width, region_start, region_end):
+        '''
+        to improve the visualization of the genes it is good to have an estimation of the label
+        length. In the following code I try to get the length of a 'W' in base pairs.
+        '''
+        if self.properties['labels'] == 'on':
+            # from http://scipy-cookbook.readthedocs.org/items/Matplotlib_LaTeX_Examples.html
+            inches_per_pt = 1.0 / 72.27
+            font_in_inches = self.properties['fontsize'] * inches_per_pt
+            region_len = region_end - region_start
+            bp_per_inch = region_len / fig_width
+            font_in_bp = font_in_inches * bp_per_inch
+            self.len_w = font_in_bp
+        else:
+            self.len_w = 1
+
+        return self.len_w
+
+    def process_bed(self):
+
+        bed_file_h = ReadBed(opener(self.properties['file']))
+        self.bed_type = bed_file_h.file_type
+
+        if 'color' in self.properties and self.properties['color'] == 'bed_rgb' and \
+           self.bed_type not in ['bed12', 'bed9']:
+            self.log.warning("*WARNING* Color set to 'bed_rgb', but bed file does not have the rgb field. The color has "
+                        "been set to {}".format(DEFAULT_BED_COLOR))
+            self.properties['color'] = DEFAULT_BED_COLOR
+
+        valid_intervals = 0
+        interval_tree = {}
+
+        max_score = float('-inf')
+        min_score = float('inf')
+        for bed in bed_file_h:
+            if bed.score < min_score:
+                min_score = bed.score
+            if bed.score > max_score:
+                max_score = bed.score
+
+            if bed.chromosome not in interval_tree:
+                interval_tree[bed.chromosome] = IntervalTree()
+
+            interval_tree[bed.chromosome].add(Interval(bed.start, bed.end, bed))
+            valid_intervals += 1
+
+        if valid_intervals == 0:
+            self.log.warning("No valid intervals were found in file {}".format(self.properties['file_name']))
+
+        return interval_tree, min_score, max_score
+
+    def get_max_num_row(self, len_w, small_relative):
+        ''' Process the whole bed regions at the given figure length and font size to
+        determine the maximum number of rows required.
+        :return:
+        '''
+
+        self.max_num_row = {}
+        for chrom in self.interval_tree:
+            row_last_position = []  # each entry in this list contains the end position
+            self.max_num_row[chrom] = 0
+            for region in sorted(self.interval_tree[chrom][0:500000000]):
+                bed = region.data
+                if self.properties['labels'] == 'on':
+                    bed_extended_end = int(bed.end + (len(bed.name) * len_w))
+                else:
+                    bed_extended_end = (bed.end + 2 * small_relative)
+
+                # get smallest free row
+                if len(row_last_position) == 0:
+                    free_row = 0
+                    row_last_position.append(bed_extended_end)
+                else:
+                    # get list of rows that are less than bed.start, then take the min
+                    idx_list = [idx for idx, value in enumerate(row_last_position) if value < bed.start]
+                    if len(idx_list):
+                        free_row = min(idx_list)
+                        row_last_position[free_row] = bed_extended_end
+                    else:
+                        free_row = len(row_last_position)
+                        row_last_position.append(bed_extended_end)
+
+                if free_row > self.max_num_row[bed.chromosome]:
+                    self.max_num_row[bed.chromosome] = free_row
+
+        self.log.debug("max number of rows set to {}".format(self.max_num_row))
+        return self.max_num_row
+
+    def get_y_pos(self, free_row):
+        """
+        The y_pos is set such that regions to be plotted do not overlap (stacked). To override this
+        the properties['collapsed'] needs to be set.
+
+        The algorithm uses a interval tree (self.region_interval) to check the overlaps
+        and a sort of coverage vector 'rows used' to identify the row in which to plot
+        :return: int y position
+        """
+
+        # if the domain directive is given, ypos simply oscilates between 0 and 100
+        if self.properties['display'] == 'interlaced':
+            ypos = self.properties['interval_height'] if self.counter % 2 == 0 else 1
+
+        elif self.properties['display'] == 'collapsed':
+            ypos = 0
+
+        else:
+            ypos = free_row * self.row_scale
+        return ypos
+
+    def plot(self, ax, label_ax, chrom_region, start_region, end_region):
+        self.counter = 0
+        self.small_relative = 0.004 * (end_region - start_region)
+        self.get_length_w(ax.get_figure().get_figwidth(), start_region, end_region)
+        if 'global max row' in self.properties and self.properties['global max row'] == 'yes':
+            self.get_max_num_row(self.len_w, self.small_relative)
+
+        if chrom_region not in self.interval_tree.keys():
+            chrom_region = self.change_chrom_names(chrom_region)
+        chrom_region = self.check_chrom_str_bytes(self.interval_tree, chrom_region)
+
+        genes_overlap = sorted(self.interval_tree[chrom_region][start_region:end_region])
+
+        # turn labels off when too many intervals are visible.
+        if self.properties['labels'] != 'off' and len(genes_overlap) > 60:
+            self.properties['labels'] = 'off'
+
+        linewidth = self.properties['line width']
+        max_num_row_local = 1
+        max_ypos = 0
+        # check for the number of other intervals that overlap
+        #    with the given interval
+        #            1         2
+        #  012345678901234567890123456
+        #  1=========       4=========
+        #       2=========
+        #         3============
+        #
+        # for 1 row_last_position = [9]
+        # for 2 row_last_position = [9, 14]
+        # for 3 row_last_position = [9, 14, 19]
+        # for 4 row_last_position = [26, 14, 19]
+
+        row_last_position = []  # each entry in this list contains the end position
+        # of genomic interval. The list index is the row
+        # in which the genomic interval was plotted.
+        # Any new genomic interval that wants to be plotted,
+        # knows the row to use by finding the list index that
+        # is larger than its start
+
+        # check for overlapping genes including
+        # label size (if plotted)
+
+        for region in genes_overlap:
+            """
+            BED12 gene format with exon locations at the end
+            chrX    20850   23076   CG17636-RA      0       -       20850   23017   0       3       946,765,64,     0,1031,2162,
+
+            BED9
+            bed with rgb at end
+            chr2L   0       70000   ID_5    0.26864549832   .       0       70000   51,160,44
+
+            BED6
+            bed without rgb
+            chr2L   0       70000   ID_5    0.26864549832   .
+            """
+            self.counter += 1
+            bed = region.data
+
+            if self.properties['labels'] == 'on':
+                num_name_characters = len(bed.name) + 2  # +2 to account for an space before and after the name
+                bed_extended_end = int(bed.end + (num_name_characters * self.len_w))
+            else:
+                bed_extended_end = (bed.end + 2 * self.small_relative)
+
+            # get smallest free row
+            if len(row_last_position) == 0:
+                free_row = 0
+                row_last_position.append(bed_extended_end)
+            else:
+                # get list of rows that are less than bed.start, then take the min
+                idx_list = [idx for idx, value in enumerate(row_last_position) if value < bed.start]
+                if len(idx_list):
+                    free_row = min(idx_list)
+                    row_last_position[free_row] = bed_extended_end
+                else:
+                    free_row = len(row_last_position)
+                    row_last_position.append(bed_extended_end)
+
+            rgb, edgecolor = self.get_rgb_and_edge_color(bed)
+
+            ypos = self.get_y_pos(free_row)
+
+            # do not plot if the maximum interval rows to plot is reached
+            if 'gene rows' in self.properties and free_row >= int(self.properties['gene rows']):
+                continue
+
+            if free_row > max_num_row_local:
+                max_num_row_local = free_row
+            if ypos > max_ypos:
+                max_ypos = ypos
+
+            if self.bed_type == 'bed12':
+                if self.properties['style'] == 'flybase':
+                    self.draw_gene_with_introns_flybase_style(ax, bed, ypos, rgb, edgecolor, linewidth)
+                else:
+                    self.draw_gene_with_introns(ax, bed, ypos, rgb, edgecolor, linewidth)
+            else:
+                self.draw_gene_simple(ax, bed, ypos, rgb, edgecolor, linewidth)
+
+            if self.properties['labels'] == 'off':
+                pass
+            elif bed.start > start_region and bed.end < end_region:
+                ax.text(bed.end + self.small_relative, ypos + (float(self.properties['interval_height']) / 2),
+                        bed.name, horizontalalignment='left',
+                        verticalalignment='center', fontproperties=self.fp)
+
+        if self.counter == 0:
+            self.log.warning("*Warning* No intervals were found for file {} "
+                             "in section '{}' for the interval plotted ({}:{}-{}).\n".
+                             format(self.properties['file'], self.properties['section_name'], chrom_region, start_region, end_region))
+        ymax = 0
+
+        if 'global max row' in self.properties and self.properties['global max row'] == 'yes':
+            ymin = self.max_num_row[chrom_region] * self.row_scale
+
+        elif 'gene rows' in self.properties:
+            ymin = int(self.properties['gene rows']) * self.row_scale
+        else:
+            ymin = max_ypos + self.properties['interval_height']
+
+        self.log.debug("ylim {},{}".format(ymin, ymax))
+        # the axis is inverted (thus, ymax < ymin)
+        ax.set_ylim(ymin, ymax)
+
+        if 'display' in self.properties:
+            if self.properties['display'] == 'domain':
+                ax.set_ylim(-5, 205)
+            elif self.properties['display'] == 'collapsed':
+                ax.set_ylim(-5, 105)
+
+        label_ax.text(0.15, 1, self.properties['title'],
+                      horizontalalignment='left', size='large',
+                      verticalalignment='top', transform=label_ax.transAxes)
+
+    def get_rgb_and_edge_color(self, bed):
+        rgb = self.properties['color']
+        edgecolor = self.properties['border color']
+
+        if self.colormap:
+            # translate value field (in the example above is 0 or 0.2686...) into a color
+            rgb = self.colormap.to_rgba(bed.score)
+
+        if self.properties['color'] == 'bed_rgb':
+            # if rgb is set in the bed line, this overrides the previously
+            # defined colormap
+            if self.bed_type in ['bed9', 'bed12'] and len(bed.rgb) == 3:
+                try:
+                    rgb = [float(x) / 255 for x in bed.rgb]
+                    if 'border color' in self.properties:
+                        edgecolor = self.properties['border color']
+                    else:
+                        edgecolor = self.properties['color']
+                except IndexError:
+                    rgb = DEFAULT_BED_COLOR
+            else:
+                rgb = DEFAULT_BED_COLOR
+        return rgb, edgecolor
+
+    def draw_gene_simple(self, ax, bed, ypos, rgb, edgecolor, linewidth):
+        """
+        draws an interval with direction (if given)
+        """
+        from matplotlib.patches import Polygon
+
+        if bed.strand not in ['+', '-']:
+            ax.add_patch(Rectangle((bed.start, ypos), bed.end - bed.start, self.properties['interval_height'],
+                                   edgecolor=edgecolor, facecolor=rgb, linewidth=linewidth))
+        else:
+            vertices = self._draw_arrow(ax, bed.start, bed.end, bed.strand, ypos)
+            ax.add_patch(Polygon(vertices, closed=True, fill=True,
+                                 edgecolor=edgecolor,
+                                 facecolor=rgb,
+                                 linewidth=linewidth))
+
+    def draw_gene_with_introns_flybase_style(self, ax, bed, ypos, rgb, edgecolor, linewidth):
+        """
+        draws a gene using different styles
+        """
+        from matplotlib.patches import Polygon
+        if bed.block_count == 0 and bed.thick_start == bed.start and bed.thick_end == bed.end:
+            self.draw_gene_simple(ax, bed, ypos, rgb, edgecolor)
+            return
+        half_height = float(self.properties['interval_height']) / 2
+        # draw 'backbone', a line from the start until the end of the gene
+        ax.plot([bed.start, bed.end], [ypos + half_height, ypos + half_height], 'black',
+                linewidth=linewidth, zorder=-1)
+
+        # get start, end of all the blocks
+        positions = []
+        for idx in range(0, bed.block_count):
+            x0 = bed.start + bed.block_starts[idx]
+            x1 = x0 + bed.block_sizes[idx]
+            if x0 < bed.thick_start < x1:
+                positions.append((x0, bed.thick_start, 'UTR'))
+                positions.append((bed.thick_start, x1, 'coding'))
+
+            elif x0 < bed.thick_end < x1:
+                positions.append((x0, bed.thick_end, 'coding'))
+                positions.append((bed.thick_end, x1, 'UTR'))
+
+            else:
+                if x1 < bed.thick_start or x0 > bed.thick_end:
+                    type = 'UTR'
+                else:
+                    type = 'coding'
+
+                positions.append((x0, x1, type))
+
+        # plot all blocks as rectangles except the last if the strand is + or
+        # the first is the strand is -, which are drawn as arrows.
+        if bed.strand == '-':
+            positions = positions[::-1]
+
+        first_pos = positions.pop()
+        if first_pos[2] == 'UTR':
+            _rgb = 'grey'
+        else:
+            _rgb = rgb
+
+        vertices = self._draw_arrow(ax, first_pos[0], first_pos[1], bed.strand, ypos)
+
+        ax.add_patch(Polygon(vertices, closed=True, fill=True,
+                             edgecolor=edgecolor,
+                             facecolor=_rgb,
+                             linewidth=linewidth))
+
+        for start_pos, end_pos, _type in positions:
+            if _type == 'UTR':
+                _rgb = 'grey'
+            else:
+                _rgb = rgb
+            vertices = [(start_pos, ypos), (start_pos, ypos + self.properties['interval_height']),
+                        (end_pos, ypos + self.properties['interval_height']), (end_pos, ypos)]
+
+            ax.add_patch(Polygon(vertices, closed=True, fill=True,
+                                 edgecolor=edgecolor,
+                                 facecolor=_rgb,
+                                 linewidth=linewidth))
+
+    def _draw_arrow(self, ax, start, end, strand, ypos):
+        """
+        Draws a filled arrow
+        :param ax:
+        :param start:
+        :param end:
+        :param strand:
+        :param ypos:
+        :param rgb:
+        :return: None
+        """
+        half_height = float(self.properties['interval_height']) / 2
+        if strand == '+':
+            x0 = start
+            x1 = end  # - self.small_relative
+            y0 = ypos
+            y1 = ypos + self.properties['interval_height']
+            """
+            The vertices correspond to 5 points along the path of a form like the following,
+            starting in the lower left corner and progressing in a clock wise manner.
+
+            -----------------\
+            ---------------- /
+
+            """
+
+            vertices = [(x0, y0), (x0, y1), (x1, y1), (x1 + self.small_relative, y0 + half_height), (x1, y0)]
+
+        else:
+            x0 = start  # + self.small_relative
+            x1 = end
+            y0 = ypos
+            y1 = ypos + self.properties['interval_height']
+            """
+            The vertices correspond to 5 points along the path of a form like the following,
+            starting in the lower left corner and progressing in a clock wise manner.
+
+            /-----------------
+            \-----------------
+            """
+            vertices = [(x0, y0), (x0 - self.small_relative, y0 + half_height), (x0, y1), (x1, y1), (x1, y0)]
+
+        return vertices
+
+    def draw_gene_with_introns(self, ax, bed, ypos, rgb, edgecolor, linewidth):
+        """
+        draws a gene like in flybase gbrowse.
+        """
+        from matplotlib.patches import Polygon
+
+        if bed.block_count == 0 and bed.thick_start == bed.start and bed.thick_end == bed.end:
+            self.draw_gene_simple(ax, bed, ypos, rgb, edgecolor, linewidth)
+            return
+        half_height = float(self.properties['interval_height']) / 2
+        quarter_height = float(self.properties['interval_height']) / 4
+        three_quarter_height = quarter_height * 3
+
+        # draw 'backbone', a line from the start until the end of the gene
+        ax.plot([bed.start, bed.end], [ypos + half_height, ypos + half_height], 'black', linewidth=linewidth, zorder=-1)
+
+        for idx in range(0, bed.block_count):
+            x0 = bed.start + bed.block_starts[idx]
+            x1 = x0 + bed.block_sizes[idx]
+            if x1 < bed.thick_start or x0 > bed.thick_end:
+                y0 = ypos + quarter_height
+                y1 = ypos + three_quarter_height
+            else:
+                y0 = ypos
+                y1 = ypos + self.properties['interval_height']
+
+            if x0 < bed.thick_start < x1:
+                vertices = ([(x0, ypos + quarter_height), (x0, ypos + three_quarter_height),
+                             (bed.thick_start, ypos + three_quarter_height),
+                             (bed.thick_start, ypos + self.properties['interval_height']),
+                             (bed.thick_start, ypos + self.properties['interval_height']),
+                             (x1, ypos + self.properties['interval_height']), (x1, ypos),
+                             (bed.thick_start, ypos), (bed.thick_start, ypos + quarter_height)])
+
+            elif x0 < bed.thick_end < x1:
+                vertices = ([(x0, ypos),
+                             (x0, ypos + self.properties['interval_height']),
+                             (bed.thick_end, ypos + self.properties['interval_height']),
+                             (bed.thick_end, ypos + three_quarter_height),
+                             (x1, ypos + three_quarter_height),
+                             (x1, ypos + quarter_height),
+                             (bed.thick_end, ypos + quarter_height),
+                             (bed.thick_end, ypos)])
+            else:
+                vertices = ([(x0, y0), (x0, y1), (x1, y1), (x1, y0)])
+
+            ax.add_patch(Polygon(vertices, closed=True, fill=True,
+                                 linewidth=linewidth,
+                                 edgecolor='none',
+                                 facecolor=rgb))
+
+            if idx < bed.block_count - 1:
+                # plot small arrows using the character '<' or '>' over the back bone
+                intron_length = bed.block_starts[idx + 1] - (bed.block_starts[idx] + bed.block_sizes[idx])
+                marker = 5 if bed.strand == '+' else 4
+                if intron_length > 3 * self.small_relative:
+                    pos = np.arange(x1 + 1 * self.small_relative,
+                                    x1 + intron_length + self.small_relative, int(2 * self.small_relative))
+                    ax.plot(pos, np.zeros(len(pos)) + ypos + half_height, '.', marker=marker,
+                            fillstyle='none', color='blue', markersize=3)
+
+                elif intron_length > self.small_relative:
+                    intron_center = x1 + int(intron_length) / 2
+                    ax.plot([intron_center], [ypos + half_height], '.', marker=5,
+                            fillstyle='none', color='blue', markersize=3)
