+++ pygenometracks/utilities.py
@@ -1,4 +1,5 @@
 import sys
+from intervaltree import IntervalTree, Interval
 
 
 def to_string(s):
@@ -29,3 +30,97 @@ def to_bytes(s):
     if isinstance(s, list):
         return [to_bytes(x) for x in s]
     return s
+
+
+def opener(filename):
+    """
+    Determines if a file is compressed or not
+    """
+    import gzip
+    f = open(filename, 'rb')
+    if f.read(2) == b'\x1f\x8b':
+        f.seek(0)
+        return gzip.GzipFile(fileobj=f)
+    else:
+        f.seek(0)
+        return f
+
+
+def file_to_intervaltree(file_name):
+    """
+    converts a BED like file into a bx python interval tree
+    :param file_name: string file name
+    :return: interval tree dictionary. They key is the chromosome/contig name and the
+    value is an IntervalTree. Each of the intervals have as 'value' the fields[3:] if any.
+    """
+    # iterate over a BED like file
+    # saving the data into an interval tree
+    # for quick retrieval
+    file_h = opener(file_name)
+    line_number = 0
+    valid_intervals = 0
+    prev_chrom = None
+    prev_start = -1
+    prev_line = None
+    interval_tree = {}
+    min_value = float('Inf')
+    max_value = -float('Inf')
+
+    for line in file_h.readlines():
+        line_number += 1
+        line = to_string(line)
+        if line.startswith('browser') or line.startswith('track') or line.startswith('#'):
+            continue
+        fields = line.strip().split('\t')
+        try:
+            chrom, start, end = fields[0:3]
+        except Exception as detail:
+            msg = "Error reading line: {}\nError message: {}".format(line_number, detail)
+            sys.exit(msg)
+
+        try:
+            start = int(start)
+        except ValueError as detail:
+            msg = "Error reading line: {}. The start field is not " \
+                  "an integer.\nError message: {}".format(line_number, detail)
+            sys.exit(msg)
+
+        try:
+            end = int(end)
+        except ValueError as detail:
+            msg = "Error reading line: {}. The end field is not " \
+                  "an integer.\nError message: {}".format(line_number, detail)
+            sys.exit(msg)
+
+        if prev_chrom == chrom:
+            assert prev_start <= start, \
+                "Bed file not sorted. Please use a sorted bed file.\n{}{} ".format(prev_line, line)
+
+        if chrom not in interval_tree:
+            interval_tree[chrom] = IntervalTree()
+
+        value = None
+
+        if len(fields) > 3:
+            value = fields[3:]
+            try:
+                line_min = min(map(float, value))
+                if line_min < min_value:
+                    min_value = line_min
+
+                line_max = max(map(float, value))
+                if line_max > max_value:
+                    max_value = line_max
+            except ValueError:
+                pass
+
+        assert end > start, "Start position larger or equal than end for line\n{} ".format(line)
+
+        interval_tree[chrom].add(Interval(start, end, value))
+        valid_intervals += 1
+
+    if valid_intervals == 0:
+        sys.stderr.write("No valid intervals were found in file {}".format(file_name))
+    file_h.close()
+
+    return interval_tree, min_value, max_value
